@@ -1,9 +1,8 @@
 module github.com/wolfeidau/shortuuid
 
-go 1.13
+go 1.23
 
 require (
-	github.com/google/uuid v1.1.1
-	github.com/mr-tron/base58 v1.1.3
-	github.com/pkg/errors v0.9.1
+	github.com/google/uuid v1.6.0
+	github.com/mr-tron/base58 v1.2.0
 )
